+++ codecov.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+set -e
+
+bash <(curl -s https://codecov.io/bash)