+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## Master
+# Next
 
+* Improved SPM integration and CI scripts
 * Added `validationClosure` properties on `APIRequest`, `DownloadAPIRequest` and `UploadAPIRequest` that allow customizing validation behaviour. Default behaviour, as before, is calling `validate()` method on `Request`.
 
 ## [3.0.1](https://github.com/MLSDev/TRON/releases/tag/3.0.1)